api/requirements.txt
azure-functions==1.21.3
azure-data-tables==12.5.0
